+++ src/Queues/MessageEnvelope.cs
@@ -28,7 +28,7 @@ namespace PipServices3.Messaging.Queues
         /// <param name="correlationId">(optional) transaction id to trace execution through call chain.</param>
         /// <param name="messageType">a string value that defines the message's type.</param>
         /// <param name="message">the data being sent/received.</param>
-        public MessageEnvelope(string correlationId, string messageType, string message)
+        public MessageEnvelope(string correlationId, string messageType, byte[] message)
         {
             CorrelationId = correlationId;
             MessageType = messageType;
@@ -43,11 +43,26 @@ namespace PipServices3.Messaging.Queues
         /// <param name="correlationId">(optional) transaction id to trace execution through call chain.</param>
         /// <param name="messageType">a string value that defines the message's type.</param>
         /// <param name="message">the data being sent/received.</param>
+        public MessageEnvelope(string correlationId, string messageType, string message)
+        {
+            CorrelationId = correlationId;
+            MessageType = messageType;
+            SetMessageAsString(message);
+            MessageId = IdGenerator.NextLong();
+        }
+
+        /// <summary>
+        /// Creates a new MessageEnvelop, which adds a correlation id, message id, and a
+        /// type to the data being sent/received.
+        /// </summary>
+        /// <param name="correlationId">(optional) transaction id to trace execution through call chain.</param>
+        /// <param name="messageType">a string value that defines the message's type.</param>
+        /// <param name="message">the data being sent/received.</param>
         public MessageEnvelope(string correlationId, string messageType, object message)
         {
             CorrelationId = correlationId;
             MessageType = messageType;
-            SetMessage(message);
+            SetMessageAsObject(message);
             MessageId = IdGenerator.NextLong();
         }
 
@@ -56,32 +71,67 @@ namespace PipServices3.Messaging.Queues
         public object Reference { get; set; }
 
         /** The unique business transaction id that is used to trace calls across components. */
-        [DataMember]
+        [DataMember(Name = "correlation_id")]
         public string CorrelationId { get; set; }
 
         /** The message's auto-generated ID. */
-        [DataMember]
+        [DataMember(Name = "message_id")]
         public string MessageId { get; set; }
 
         /** String value that defines the stored message's type. */
-        [DataMember]
+        [DataMember(Name = "message_type")]
         public string MessageType { get; set; }
 
         /** The time at which the message was sent. */
-        [DataMember]
+        [DataMember(Name = "sent_time")]
         public DateTime SentTime { get; set; }
 
         /** The stored message. */
-        [DataMember]
-        public string Message { get; set; }
+        [IgnoreDataMember]
+        public byte[] Message { get; set; }
+
+        /** Used for serialization */
+        [DataMember(Name = "message")]
+        public string MessageBase64
+        {
+            get
+            {
+                if (Message == null)
+                {
+                    return null;
+                }
+                else
+                {
+                    return Convert.ToBase64String(Message);
+                }
+            }
+            set
+            {
+                if (value == null)
+                {
+                    Message = null;
+                }
+                else
+                {
+                    Message = Convert.FromBase64String(value);
+                }
+            }
+        }
 
         /// <summary>
-        /// Stores the given value as a JSON string.
+        /// Stores the given value as a string.
         /// </summary>
-        /// <param name="message">the value to convert to JSON and store in this message.</param>
-        public void SetMessage(object message)
+        /// <param name="message">A string value for this message.</param>
+        public void SetMessageAsString(string message)
         {
-            Message = JsonConverter.ToJson(message);
+            if (message == null)
+            {
+                Message = null;
+            }
+            else
+            {
+                Message = Encoding.UTF8.GetBytes(message);
+            }
         }
 
         /// <summary>
@@ -89,9 +139,52 @@ namespace PipServices3.Messaging.Queues
         /// </summary>
         /// <typeparam name="T">the class type</typeparam>
         /// <returns>the value that was stored in this message as a JSON string.</returns>
-        public T GetMessage<T>()
+        public string GetMessageAsString()
         {
-            return JsonConverter.FromJson<T>(Message);
+            if (this.Message == null)
+            {
+                return null;
+            }
+            else
+            {
+                return Encoding.UTF8.GetString(this.Message);
+            }
+        }
+
+        /// <summary>
+        /// Stores the given value as a JSON string.
+        /// </summary>
+        /// <param name="message">the value to convert to JSON and store in this message.</param>
+        public void SetMessageAsObject(object message)
+        {
+            if (message == null)
+            {
+                Message = null;
+            }
+            else
+            {
+                var json = JsonConverter.ToJson(message);
+                Message = Encoding.UTF8.GetBytes(json);
+            }
+
+        }
+
+        /// <summary>
+        /// Gets the value that was stored in this message as a JSON string.
+        /// </summary>
+        /// <typeparam name="T">the class type</typeparam>
+        /// <returns>the value that was stored in this message as a JSON string.</returns>
+        public T GetMessageAs<T>()
+        {
+            if (this.Message == null)
+            {
+                return default(T);
+            }
+            else
+            {
+                var json = Encoding.UTF8.GetString(this.Message);
+                return JsonConverter.FromJson<T>(json);
+            }
         }
 
         /// <summary>
@@ -110,7 +203,7 @@ namespace PipServices3.Messaging.Queues
             builder.Append(',');
             builder.Append(MessageType ?? "---");
             builder.Append(',');
-            var sample = Message ?? "---";
+            var sample = GetMessageAsString() ?? "---";
             sample = sample.Length > 150 ? sample.Substring(0, 150) : sample;
             builder.Append(sample);
             builder.Append(']');